Adding files to cmake, and using the feature toggle class
