+++ src/grpc_client.cc
@@ -10,6 +10,7 @@
 #include <grpcpp/support/channel_arguments.h>
 #include <ctime>
 #include <chrono>
+#include "feature_toggles.h"
 
 namespace grpc_labview
 {
@@ -276,6 +277,9 @@ LIBRARY_EXPORT int32_t ClientUnaryCall(
     int32_t timeoutMs,
     grpc_labview::gRPCid* contextId)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto client = clientId->CastTo<grpc_labview::LabVIEWgRPCClient>();
     if (!client)
     {
@@ -401,6 +405,9 @@ LIBRARY_EXPORT int32_t ClientBeginClientStreamingCall(
     int32_t timeoutMs,
     grpc_labview::gRPCid* contextId)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto client = clientId->CastTo<grpc_labview::LabVIEWgRPCClient>();
     if (!client)
     {
@@ -455,6 +462,9 @@ LIBRARY_EXPORT int32_t ClientBeginServerStreamingCall(
     int32_t timeoutMs,
     grpc_labview::gRPCid* contextId)
 {    
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto client = clientId->CastTo<grpc_labview::LabVIEWgRPCClient>();
     if (!client)
     {
@@ -517,6 +527,9 @@ LIBRARY_EXPORT int32_t ClientBeginBidiStreamingCall(
     int32_t timeoutMs,
     grpc_labview::gRPCid* contextId)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto client = clientId->CastTo<grpc_labview::LabVIEWgRPCClient>();
     if (!client)
     {
@@ -563,6 +576,9 @@ LIBRARY_EXPORT int32_t ClientBeginBidiStreamingCall(
 //---------------------------------------------------------------------
 LIBRARY_EXPORT int32_t ClientBeginReadFromStream(grpc_labview::gRPCid* callId, grpc_labview::MagicCookie* occurrencePtr)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto reader = callId->CastTo<grpc_labview::StreamReader>();
     auto call = callId->CastTo<grpc_labview::ClientCall>();
 
@@ -592,6 +608,9 @@ LIBRARY_EXPORT int32_t ClientBeginReadFromStream(grpc_labview::gRPCid* callId, g
 //---------------------------------------------------------------------
 LIBRARY_EXPORT int32_t ClientCompleteReadFromStream(grpc_labview::gRPCid* callId, int* success, int8_t* responseCluster)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+
     auto reader = callId->CastTo<grpc_labview::StreamReader>();
     auto call = callId->CastTo<grpc_labview::ClientCall>();
     if (!reader || !call)
@@ -618,6 +637,9 @@ LIBRARY_EXPORT int32_t ClientCompleteReadFromStream(grpc_labview::gRPCid* callId
 //---------------------------------------------------------------------
 LIBRARY_EXPORT int32_t ClientWriteToStream(grpc_labview::gRPCid* callId, int8_t* requestCluster, int* success)
 {
+    // Instantiating the feature toggles singleton that will read the feature configuration file
+    FeatureConfig::getInstance().readConfigFromFile("feature_config.ini");
+    
     auto writer = callId->CastTo<grpc_labview::StreamWriter>();
     if (!writer)
     {
